requirements/local.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-model-utils==4.2.0
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1